Changes to Structure, Progress on Graph-Building
-Converted the TilePuzzleModel to store the tiles as nodes from the
start (And updated functions dealing with the user editing the colors of
each tile) rather than storing the colors as integers, and as nodes, for
simplicity. May even be quicker if I can get it to update edges as
single tiles are edited rather than constantly rebuilding the whole
graph, uncertain if that will happen. Need to further edit the Node
class to be able to hold "infinite" edges (probably an arraylist)
because the start and end nodes can have more than 4 edges.

-Started writing code for determining the edges between the nodes to
build a graph to feed to the a*-based path-finding algorithm. About 2/3
done, then I want to write code to test this thoroughly, at the very
least printing out all edges, but hopefully graphically displaying each
edge between nodes (will make it much easier to test). Also need to come
up with test case puzzle(s) and maybe a way to load those puzzles each
time?