Graph Building mostly coded
-checkEdgeDown(node, int, int) was written/modified from the code for
checkEdgeLeft(node, int, int).

-Mistakes in checkEdgeLeft were found while writing checkEdgeDown and
were corrected.

- Quickly wrote an event for the solve button to run buildGraph() to see
what kind of performance hit building the whole graph at once had versus
dynamically building the graph as the puzzle is edited.

No significant performance usage was measured with the current default
puzzle size/configuration and in buildGraph()'s current somewhat
unfinished state.

Next To Do:

-Still need to edit the node structure to be able to accept more that 4
nodes (for start and end nodes).

The dilemma is that directions of edges are supposed to be fixed to an
index in an array (size 4) so I can also keep track if passing over an
edge changes the "scent". This is primarily an issue when an edge
"passes over" a purple tile due to the rules of the puzzle; the player
gets the lemon scent but doesn't end up on the purple tile, so we can't
just check if the tile is purple to determine if the scent is lemon.

- Once the above (significant) issue is fixed, need to write code to
display edges that were determined from buildGraph() for testing
purposes, as described in the last commit.