Rewrite/Reorganization of buildGraph-related methods
- Replaced testEdgeDown and testEdgeLeft (which should have been right
but I was not smart) with findEdgesForNode and checkForEdgeInDirection.

This replaces the disorganized mess with a much more organized system
that was modeled after a finite state machine, and while
testForEdgeInDirection is less specific and therefore somewhat more
complicated, the benefit is the decreased risk of mistakes when making
changes in either checkEdgeLeft or Down, and forgetting to bring it to
the other or adapting it incorrectly.

-Updated buildGraph to use these new methods.

-Modified addEdges to addEdge, only adds edge for the one specified
direction.

This means it has to be used twice in most cases, but it solves the
bug/problem of edges being added in both directions when they shouldn't
be in some special cases involving purple sliding tiles without having
to write some complex tests for these situations.
---------------------------------
The rewritten methods don't currently place edges correctly in many
situations and is less correct than what it aims to replace so far, and
is in no way ready to be merged to the master branch.