Testing & Bug Fixes
-Identified and fixed a bug in edge checking involving purple tiles
where the color of the possible next tile was being checked before it
was checked if it'd even be within the bounds.
-Fixed a rather silly bug where rightwards and upwards dummy nodes/edges
weren't being made (switched up dy and dx, so it was failing checks).
-Identified & fixed a bug involving dummy nodes being made when they
shouldn't (say, if it'd slide into another yellow tile)/being made but
the edge going from the dummy node would go back to the wrong tile if
nodeBeingChecked was purple.
- Expanded the edgesToString() method in Node to give information about
what tile a dummy node will send a player back to (related to prev. bug)