GUI Features
-Implemented quick features to allow for displaying the graph edges,
then editing the puzzle or loading another (removing displayed edges)
and then displaying the graph again when the graph button is pressed,
rather than having to restart the application
--------------------------------------------------
NEXT:
-I am fairly happy with the status of the buildGraph-related functions,
and it's functioned better than the old version in the master branch for
a long time, a merge is long overdue and will be done next.
-There is not much more testing I have come up with to do right now, so
I'm fairly confident most/all the major bugs are gone.
-Move on to the path-finding algorithm.