Scented A* (DOES NOT WORK YET)
- Modified the A* pathfinding algorithm to consider the scents mechanic
of this particular puzzle. However, it currently seems to not work
because it allows for going over the same tile multiple times, and the
more traditional way of storing the path I've chosen (storing the step
before it as a parent Node) doesn't work for this application. If we go
back to a tile that's already been checked in an orange scent loop, we
overwrite its parent node information from when it was in said loop,
which potentially erases part of the path.