Partial Bug Fix of Path Breaking
- Fixed one instance of the bug where the tree that stores the path
breaks when becoming non-orange scented (that is, the step where it
becomes non-orange scented has its parent set to null) where a pink tile
above a purple tile that was against the tile's lower bounds in
dummyNodeErrorSmallTest.txt had an edge that wasn't lemon-scented.

Because of this, the orange-scented loop continued to test on the purple
space, removing it from the list of leaves of the path tree. Then, when
the purple space was tested while not orange-scented, it wasn't in the
leaves list, so currentStep was null, so the next step's parent was
null, breaking the path.

To fix this, a bug was fixed in the checkForEdgeInDirection method in
case 6.
------------------------------------
Similar bugs still cause the path to break in spots where the path
switches to being non-orange scented in other scenarios, like the
default puzzle after a dummy node. The cause is likely similar to the
bug that was just fixed. Further investigation/testing needed.