Bug fix
-Added an isOrangeStep bool to PathTreeNode to keep track of whether or
not a step was made while in an orange loop
- Fixed a bug where a null PathTreeNode could make its way into Leaves
and crash the program.

This happened when a step onto a certain space was previously made, and
a step is made onto that space again in an orange loop this time, but
the previous step had a lower height. With no way of knowing this lower
height leaf did not take place in the orange loop, the algorithm selects
it instead and removes all leaves for that position. Later, when that
space is dequeued from the openSet, there is no leaf for it, so the
currentStep = null.
-------------------------
NEXT TO DO:
- Search for more bugs, design test cases
- Fix graphical bug(s) in drawing of path
- Implement avoid green tiles mode
- Clean up code