Node/Edge structure modification
-Changed the Node class to store its row/column position, because
previously had to do a bit of a hacky workaround to access a node's
position.
-Changed Edge class so it no longer has to hold rows/columns of
parent/child nodes.
-Updated all places in the code accordingly to reflect these changes.
-Had to add extra checks in the solution algorithm related to the
start/end nodes as a result of the changes (because start/end now have a
specific row/column, was throwing off f/g calculations to favor tiles in
the same row as the start/end node is set to).