Avoid Green Tiles Implemented
-Implemented the avoid green tiles function that, when the appropriate
checkbox is checked, the pathfinding algorithm will try to find the
shortest path that avoids going over green tiles.
-Added a weight variable to the Node class which is zero by default, but
when the avoid green tiles check box is checked, all green tiles are
given a weight equal to the number of columns
-Changed resize puzzle functions to properly handle/update weight
-Changed the pathfinding algorithm to take into account weight when
determining a tile's g value, so tiles with more weight will be avoided.