resizeTileGrid Optimization & More Code Cleanup
-Realized the TilePuzzleModel 2d array wouldn't need to be resized if we
encounter a new size too large for memory (code from last commit) and
removed the 2nd call to the resizePuzzle method.
-Removed unnecessary parameters from the constructor of an Edge, and
from all methods dealing with adding an Edge between two tiles that I
forgot to remove previously when edges were made to only need to store a
reference to the node they go to, not the positions of each node
involved.