UI Enhancements & More Help Info
-Added tabs for different sections of info in the help window and added
sections on how to use the program.
-Made it so that clicking a tile while not in edit mode before the
puzzle has been graphed does not pop up a blank window by adding the
isGraphed variable to mainwindow.xaml.cs
-Added a "no edges" message to the window that shows all edges of a tile
when the puzzle is graphed if the tile has no edges (rather than showing
a blank window)
-Made it so the button of the selected color has a thicker, white
border, to let the user know which is selected. Added
unselectColorButtons() to mainwindow.xaml.cs which resets the look of
each color buttons' border before the button click event applies the
selected appearance to a specific button.